requirements.txt
matminer==0.9.2
pymatgen
streamlit
pandas
numpy
scipy
scikit-learn
matplotlib
mysql-connector-python
joblib
ipykernel